backend/services/admin/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
PyJWT>=2.8.0
passlib[bcrypt]>=1.7.4
redis>=5.0.0
python-multipart>=0.0.6
